+++ transforms/code/proglang_select/python/.dockerignore
@@ -1,0 +1,1 @@
+venv/
