+++ transforms/code/proglang_select/python/test-data/expected/metadata.json
@@ -1,0 +1,52 @@
+{
+    "pipeline": "pipeline_id",
+    "job details": {
+        "job category": "preprocessing",
+        "job name": "lang_select",
+        "job type": "ray",
+        "job id": "job_id",
+        "start_time": "2024-03-12 07:45:26",
+        "end_time": "2024-03-12 07:45:28",
+        "status": "success"
+    },
+    "code": {
+        "github": "github",
+        "commit_hash": "12345",
+        "path": "path"
+    },
+    "job_input_params": {
+        "ls_language_column": "language",
+        "ls_allowed_langs_file": "/Users/boris/Projects/fm-data-engineering/transforms/code/select_language/test-data/languages/allowed-code-languages.txt",
+        "ls_return_known": true,
+        "checkpointing": false,
+        "max_files": -1,
+        "number of workers": 1,
+        "worker options": {
+            "num_cpus": 0.8
+        },
+        "actor creation delay": 0
+    },
+    "execution_stats": {
+        "cpus": 16,
+        "gpus": 0,
+        "memory": 11.76053390558809,
+        "object_store": 2.0
+    },
+    "job_output_stats": {
+        "source_files": 1,
+        "source_size": 31,
+        "result_files": 1,
+        "result_size": 32,
+        "table_processing": 0.017200946807861328,
+        "documents with supported languages": 2,
+        "documents with unsupported languages": 0
+    },
+    "source": {
+        "name": "/Users/boris/Projects/fm-data-engineering/transforms/code/select_language/test-data/input",
+        "type": "path"
+    },
+    "target": {
+        "name": "/Users/boris/Projects/fm-data-engineering/transforms/code/select_language/test-data/output",
+        "type": "path"
+    }
+}
