+++ transforms/code/proglang_select/ray/Dockerfile
@@ -9,13 +9,19 @@ COPY --chown=ray:users data-processing-lib-python/ data-processing-lib-python/
 RUN cd data-processing-lib-python && pip install --no-cache-dir -e .
 COPY --chown=ray:users data-processing-lib-ray/ data-processing-lib-ray/ 
 RUN cd data-processing-lib-ray    && pip install --no-cache-dir -e .
+COPY --chown=ray:users python-transform/  python-transform/
+RUN cd python-transform && pip install --no-cache-dir -e .
 
-COPY requirements.txt requirements.txt
-RUN pip install --no-cache-dir -r  requirements.txt
+#COPY requirements.txt requirements.txt
+#RUN pip install --no-cache-dir -r  requirements.txt
+
+COPY --chown=ray:users src/ src/
+COPY --chown=ray:users pyproject.toml pyproject.toml 
+RUN pip install --no-cache-dir -e .
 
 # copy source data
-COPY src/proglang_select_transform_ray.py .
-COPY src/proglang_select_local_ray.py local/
+COPY ./src/proglang_select_transform_ray.py .
+COPY ./src/proglang_select_local_ray.py local/
 
 # copy test
 COPY test/ test/
