--- transforms/code/proglang_select/ray/requirements.txt
@@ -1,2 +1,0 @@
-# transform runtime
-#data-prep-kit==0.1.2
